Added a Result variable to playRound() function
I have added functionality so that, as well as producing a
console log statement, each switch case in playRound() also
logs a variable called result. This variable will contain the
values: win, lose or draw. The purpose behind this is to use
this variable for keeping score in the game() function.